Day 13, part 2 incomplete
One remaining issue, pattern 4 returning 100 rather than 5

+++ 2023/Day13.h
@@ -7,7 +7,7 @@ private:
 	const std::string testFile2 = "inputs/day13_part2_test.txt";
 	const std::string puzzleFile = "inputs/day13.txt";
 	int calculatePuzzle1(std::vector<std::string> input);
-	int calculatePuzzle2(std::vector<std::string> input);
+	int64_t calculatePuzzle2(std::vector<std::string> input);
 public:
 	void puzzle1();
 	void puzzle2();
